requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.10.1
apache-airflow-providers-postgres>=5.0.0
astronomer-cosmos[dbt-postgres]